Init Dynamics header split
