+++ src/dsm/dsm.hpp
@@ -25,7 +25,7 @@ namespace dsm {
 #include "headers/Roundabout.hpp"
 #include "headers/SparseMatrix.hpp"
 #include "headers/Street.hpp"
-#include "headers/Dynamics.hpp"
+#include "headers/FirstOrderDynamics.hpp"
 #include "utility/TypeTraits/is_node.hpp"
 #include "utility/TypeTraits/is_street.hpp"
 #include "utility/TypeTraits/is_numeric.hpp"
